backend/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database and ORM
sqlalchemy==2.0.36

# Data validation and settings
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.3.0

# File handling
python-multipart==0.0.20

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Video/Image processing
opencv-python-headless==4.12.0.88
numpy==2.0.2
mediapipe==0.10.14

# Audio processing
ffmpeg-python==0.2.0
openai-whisper==20250625
torch==2.5.1
torchaudio==2.5.1
soundfile==0.13.1

# PDF Processing
PyPDF2==3.0.1

# LLM API
google-generativeai==0.8.5

# GitHub API & HTTP Requests
requests==2.32.3
aiohttp==3.11.11

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.3.0

# Additional dependencies
scipy==1.13.1
